Framework and slickgrid updates.

+++ app/models/organization.rb
@@ -1,36 +1,36 @@
-class Organization < ActiveRecord::Base 
-	
-#	===========================
-# 	Association declarations:
-#	===========================
- 
-	belongs_to :party,:dependent => :destroy
-	has_many :facilities
+class Organization < ActiveRecord::Base
+
+  #  ===========================
+  #   Association declarations:
+  #  ===========================
+
+  belongs_to :party,:dependent => :destroy
+  has_many :facilities
   has_many :organization_rules, :dependent => :destroy
- 
-#	============================
-#	 Validations declarations:
-#	============================
-	validates_presence_of :short_description
-#	=====================
-#	 Complex validations:
-#	=====================
-def validate 
 
-	 if self.new_record? 
-		 validate_uniqueness
-	 end
-end
+  #  ============================
+  #   Validations declarations:
+  #  ============================
+  validates_presence_of :short_description
+  #  =====================
+  #   Complex validations:
+  #  =====================
+  def validate
 
-def validate_uniqueness
-	 exists = Organization.find_by_short_description(self.short_description)
-	 if exists != nil 
-		errors.add_to_base("There already exists a record with the combined values of fields: 'short_description' ")
-	end
-end
+    if self.new_record?
+      validate_uniqueness
+    end
+  end
+
+  def validate_uniqueness
+    exists = Organization.find_by_short_description(self.short_description)
+    if exists != nil
+      errors.add_to_base("There already exists a record with the combined values of fields: 'short_description' ")
+    end
+  end
 
 
-def before_save
+  def before_save
     party = nil
     if self.new_record?
       party = Party.new
@@ -42,18 +42,18 @@ def before_save
     party.party_name = self.short_description
     party.save
     self.party = party
- 
- 
- end
 
-  def after_destroy
-    self.party.destroy
-  
+
   end
 
-    
-def Organization.get_all_by_role(role,return_records = nil)
-  
+  def after_destroy
+    self.party.destroy if self.party
+
+  end
+
+
+  def Organization.get_all_by_role(role,return_records = nil)
+
     query = "SELECT  public.organizations.id,
              public.organizations.short_description
              FROM
@@ -61,40 +61,40 @@ def Organization.get_all_by_role(role,return_records = nil)
              INNER JOIN public.roles ON (public.parties_roles.role_id = public.roles.id)
              INNER JOIN public.organizations ON (public.parties_roles.party_id = public.organizations.party_id)
              WHERE
-             (public.roles.role_name = '#{role}')"  
-  
-   orgs = Organization.find_by_sql(query)
-   if return_records
-    return orgs
-   else
-    return orgs.map{|o| [o.short_description]}
-   end
-   
+             (public.roles.role_name = '#{role}')"
+
+    orgs = Organization.find_by_sql(query)
+    if return_records
+      return orgs
+    else
+      return orgs.map{|o| [o.short_description]}
+    end
+
   end
- 
+
   def Organization.get_sell_bys_by_orgs(org1,org2)
-  
+
     query = "SELECT  public.organizations.sell_by_description
              FROM
              public.organizations
              WHERE
              ( public.organizations.short_description = '#{org1}'
-              OR public.organizations.short_description = '#{org2}')"  
+              OR public.organizations.short_description = '#{org2}')"
 
-   return Organization.find_by_sql(query).map{|o| [o.sell_by_description]}
+    return Organization.find_by_sql(query).map{|o| [o.sell_by_description]}
   end
-  
- def Organization.get_sell_bys_by_org(role,org_name)
-  
+
+  def Organization.get_sell_bys_by_org(role,org_name)
+
     query = "SELECT  public.organizations.sell_by_description
              FROM
              public.parties_roles
              INNER JOIN public.roles ON (public.parties_roles.role_id = public.roles.id)
              INNER JOIN public.organizations ON (public.parties_roles.party_id = public.organizations.party_id)
              WHERE
-             (public.roles.role_name = '#{role}') AND public.organizations.short_description = '#{org_name}'"  
+             (public.roles.role_name = '#{role}') AND public.organizations.short_description = '#{org_name}'"
 
-   return Organization.find_by_sql(query).map{|o| [o.sell_by_description]}
+    return Organization.find_by_sql(query).map{|o| [o.sell_by_description]}
   end
 
   def need_gtin_check?
@@ -109,6 +109,12 @@ def Organization.get_all_by_role(role,return_records = nil)
       end
     end
     return false
+  end
+
+  # Get an Organization from a PartyRole or a party_role_id
+  def self.find_using_party_role(party_role_or_id)
+    party_role = party_role_or_id.respond_to?(:party_id) ? party_role_or_id : PartiesRole.find(party_role_or_id)
+    Organization.find_by_party_id(party_role.party_id)
   end
 
 end
